refactoring DraftingPage
